# contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Shadowcook API
  version: 0.1.0
paths:
...
  /admin/tenants:
    get:
      operationId: listTenants
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: Tenant list. } }
    post:
      operationId: createTenant
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [firstName, lastName, email, cookbookName, slug],
                    properties:
                      {
                        firstName: { type: string },
                        lastName: { type: string },
                        email: { type: string, format: email },
                        cookbookName: { type: string },
                        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' },
                      },
                  },
              },
          }
      responses:
        {
          '201': { description: Tenant and owner invitation created. },
          '400':
            {
              description: Invalid tenant,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '409':
            {
              description: Tenant slug conflict,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '429':
            {
              description: Invitation-code request limit reached; no tenant was created,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '502':
            {
              description: SMTP delivery failed; no tenant was created,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '503':
            {
              description: SMTP is not configured; no tenant was created,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
  /admin/tenants/{publicId}:
    patch:
      operationId: updateTenant
      summary: Renames a tenant or changes its enabled state.
      security: [{ sessionCookie: [] }]
      parameters:
        [{ name: publicId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    additionalProperties: false,
                    properties:
                      { displayName: { type: string, minLength: 1 }, disabled: { type: boolean } },
                    minProperties: 1,
                  },
              },
          }
      responses:
        {
          '204': { description: Tenant updated. },
          '400':
            {
              description: Invalid update,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '404':
            {
              description: Tenant not found,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
    delete:
      operationId: deleteTenant
      summary: Permanently deletes a tenant and all tenant-owned data.
      security: [{ sessionCookie: [] }]
      parameters:
        [{ name: publicId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        {
          '204': { description: Tenant and tenant-owned data deleted. },
          '404':
            {
              description: Tenant not found,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
  /admin/users:
    get:
      operationId: listUsers
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: Users and tenant memberships. } }
  /invitations/{token}:
    get:
      operationId: getTenantInvitation
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses:
        {
          '200': { description: Invitation details. },
          '404':
            {
              description: Invitation unavailable,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
  /invitations/{token}/accept:
    post:
      operationId: acceptTenantInvitation
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [code],
                    properties: { code: { type: string, pattern: '^[0-9]{6}$' } },
                  },
              },
          }
      responses:
        {
          '204': { description: Invitation accepted. },
          '401':
            {
              description: Invalid email code,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '403':
            {
              description: The active session uses a different email address,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '404':
            {
              description: Invitation unavailable,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
...
components:
...
  schemas:
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, error]
      properties:
        code:
          type: string
          enum:
            - AUTHENTICATION_FAILED
            - AUTHENTICATION_REQUIRED
            - INVALID_CREDENTIALS
            - INVALID_LOGIN_BODY
            - INVALID_PASSWORD
            - PASSWORD_CHANGE_REQUIRED
            - RECIPE_NOT_FOUND
            - RECIPE_CONFLICT
            - INVALID_RECIPE
            - INVALID_RECIPE_CATEGORY
            - PUBLISH_CATEGORY_REQUIRED
            - RECIPE_NOT_PUBLISHED
            - INVALID_UNIT
            - UNIT_CONFLICT
            - UNIT_NOT_FOUND
            - UNIT_IN_USE
            - INVALID_CONVERSION
            - UNIT_DIMENSION_MISMATCH
            - INVALID_INGREDIENT
            - INGREDIENT_CONFLICT
            - INGREDIENT_NOT_FOUND
            - INGREDIENT_IN_USE
            - INVALID_ALIAS
            - ALIAS_CONFLICT
            - ALIAS_NOT_FOUND
        error:
          type: string
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shadowcook_session